[c# epoxy] Clean up unstarted EpoxyConnections
If an EpoxyListener or EpoxyTransport is shutdown between creation of a
EpoxyConnection, but before the connection was started, calling
StopAsync() on the connection would never complete, as the connection
loop needed to run to transition the connection to Disconnected.

Now, EpoxyConnection.StopAsync() can be called before the connection has
been started to clean up the network resources.