[New] Remove the need of BondManualAssignment annotation.
The annotation was added to express a semantic, that a field with
generic type must be manually assigned a value by developers before
serialization. However, the concept is not intuitive, especially when
combining with nullable.

After study I found a solution, that can really detect if a given field
is generic. So we don't really need BondManualAssignment annotation
anymore, since we can implicitly enforce this check by checking the type
of each field.